requirements.txt
supabase==2.13.0
groq==0.18.0
anthropic==0.49.0
openai==2.30.0
requests==2.32.3
python-dotenv==1.0.1
google-auth==2.38.0
google-auth-oauthlib==1.2.4
google-api-python-client==2.190.0
bcrypt==5.0.0
playwright==1.58.0
